+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2025-05-26
+
+### Added
+
+- You can now see a distribution chart by quartiles, and hover to see the total number of attendees at that quartile
+
 ## 2025-05-17
 
 ### Modified